+++ src/core_modules/capture-core/components/DataEntries/SingleEventRegistrationEntry/SingleEventRegistrationEntry.component.tsx
@@ -1,13 +1,21 @@
 import React from 'react';
+import { useDispatch } from 'react-redux';
+import i18n from '@dhis2/d2-i18n';
+import { NoWriteAccessMessage } from '../../NoWriteAccessMessage';
 import { NewEventDataEntryWrapper } from './DataEntryWrapper/NewEventDataEntryWrapper.container';
 import { NewRelationshipWrapper } from './NewRelationshipWrapper/NewEventNewRelationshipWrapper.container';
-import { SelectionsNoAccess } from './SelectionsNoAccess/dataEntrySelectionsNoAccess.container';
+import { cancelNewEventAndReturnToMainPage } from './DataEntryWrapper/DataEntry/actions/dataEntry.actions';
 import type { Props } from './SingleEventRegistrationEntry.types';
 
 export const SingleEventRegistrationEntryComponent = ({ showAddRelationship, eventAccess }: Props) => {
+    const dispatch = useDispatch();
+
     if (!eventAccess.write) {
         return (
-            <SelectionsNoAccess />
+            <NoWriteAccessMessage
+                message={i18n.t("You don't have access to create an event in the current selections")}
+                onCancel={() => dispatch(cancelNewEventAndReturnToMainPage())}
+            />
         );
     }
