+++ src/core_modules/capture-core/components/NoWriteAccessMessage/NoWriteAccessMessage.component.tsx
@@ -1,36 +1,36 @@
 import React, { type ComponentType } from 'react';
+import i18n from '@dhis2/d2-i18n';
+import { Button } from '@dhis2/ui';
 import { withStyles, type WithStyles } from 'capture-core-utils/styles';
 import { IncompleteSelectionsMessage } from '../IncompleteSelectionsMessage';
 
 const styles = () => ({
-    header: {
-        flexGrow: 1,
-        fontSize: 16,
-        fontWeight: 500,
-    },
     message: {
         marginTop: 10,
     },
 });
 
 type Props = {
-    title?: string;
     message: string;
+    onCancel?: () => void;
 };
 
 type PropsWithStyles = Props & WithStyles<typeof styles>;
 
-export const NoWriteAccessMessagePlain = ({ title, message, classes }: PropsWithStyles) => (
-    <>
-        <div className={classes.header}>
-            {title}
-        </div>
-        <div className={classes.message}>
-            <IncompleteSelectionsMessage>
-                {message}
-            </IncompleteSelectionsMessage>
-        </div>
-    </>
+const NoWriteAccessMessagePlain = ({ message, onCancel, classes }: PropsWithStyles) => (
+    <div className={classes.message}>
+        <IncompleteSelectionsMessage>
+            {message}
+        </IncompleteSelectionsMessage>
+        {onCancel && (
+            <Button
+                dataTest="no-write-access-cancel-button"
+                onClick={onCancel}
+            >
+                {i18n.t('Cancel')}
+            </Button>
+        )}
+    </div>
 );
 
 export const NoWriteAccessMessage =
