+++ src/core_modules/capture-core/components/Pages/New/NewPage.component.tsx
@@ -82,6 +82,7 @@ const NewPagePlain = ({
                                     interpolation: { escapeValue: false },
                                 },
                             )}
+                        onCancel={handleMainPageNavigation}
                     />
                     :
                     <OrgUnitFetcher orgUnitId={orgUnitId}>
