--- src/core_modules/capture-core/components/WidgetEnrollmentEventNew/AccessVerification/accessVerification.types.ts
@@ -1,26 +1,0 @@
-import type { ProgramStage, RenderFoundation, TrackerProgram } from '../../../metaData';
-import type { ExternalSaveHandler, RulesExecutionDependencies } from '../common.types';
-
-export type ContainerProps = {
-    program: TrackerProgram;
-    stage: ProgramStage;
-    formFoundation: RenderFoundation;
-    teiId: string;
-    enrollmentId: string;
-    orgUnitId: string;
-    rulesExecutionDependencies: RulesExecutionDependencies;
-    onSaveExternal?: ExternalSaveHandler;
-    onSaveSuccessActionType?: string;
-    onSaveErrorActionType?: string;
-    onSaveAndCompleteEnrollmentSuccessActionType?: string;
-    onSaveAndCompleteEnrollmentErrorActionType?: string;
-    widgetReducerName: string;
-    onCancel?: () => void;
-};
-
-export type Props = {
-    eventAccess: {
-        read: boolean;
-        write: boolean;
-    };
-} & ContainerProps;
