--- src/core_modules/capture-core/components/WidgetEnrollmentEventNew/AccessVerification/index.ts
@@ -1,1 +1,0 @@
-export { AccessVerification } from './AccessVerification.container';
