+++ src/core_modules/capture-core/components/WidgetEnrollmentEventNew/WidgetEnrollmentEventNew.container.tsx
@@ -1,7 +1,7 @@
 import React, { useMemo } from 'react';
 import i18n from '@dhis2/d2-i18n';
 import { getProgramAndStageForProgram, TrackerProgram } from '../../metaData';
-import { AccessVerification } from './AccessVerification';
+import { OrgUnitFetcher } from './OrgUnitFetcher/OrgUnitFetcher.container';
 import type { WidgetProps } from './WidgetEnrollmentEventNew.types';
 import { useMetadataForProgramStage } from '../DataEntries/common/ProgramStage/useMetadataForProgramStage';
 
@@ -36,7 +36,7 @@ export const WidgetEnrollmentEventNew = ({
     }
 
     return (
-        <AccessVerification
+        <OrgUnitFetcher
             {...passOnProps}
             stage={stage}
             formFoundation={formFoundation}
