--- src/core_modules/capture-core/components/WidgetEventSchedule/AccessVerification/index.ts
@@ -1,1 +1,0 @@
-export { NoAccess } from './NoAccess.component';
